LayoutTests/fast/repaint/block-inputrange-repaint.html

<!DOCTYPE html>
<head>
<title>This tests when input=range is block we issue correct repaint rects</title>
<script>jsTestIsAsync = true;</script>
<script src="../../resources/js-test-pre.js"></script>
<style>
  input[type="range"] {
    display: block;
    width: 500px;
  }
</style>
</head>
<body>
<input id=moveme type=range>
<script>
  window.internals.startTrackingRepaints();
  document.getElementById("moveme").value = 100;
  document.body.offsetHeight;
  function repaintTest() {
    if (window.internals) {
      repaintRects = window.internals.repaintRectsAsText();
      window.internals.stopTrackingRepaints();
      shouldNotBe("repaintRects.indexOf('495 8 15 15')", "-1");
    }
    finishJSTest();
  }
  setTimeout(repaintTest, 0);
</script>
</body>
<script src="../../resources/js-test-post.js"></script>
</html>
