Safari-611.1.10/LayoutTests/fast/repaint/block-inputrange-repaint-expected.txt

PASS repaintRects.indexOf('495 8 15 15') is not -1
PASS successfullyParsed is true

TEST COMPLETE
